services/backend/pong_service/requirements.txt
Django==4.2.9
channels==4.0.0
daphne==4.0.0
Twisted[tls,http2]
psycopg2-binary==2.9.9
djangorestframework==3.12.5
django-cors-headers==3.8.0

